requirements.txt
flask>=3.0.0
google-genai>=1.0.0
gunicorn>=21.0.0
httpx>=0.27.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.0.0
fpdf2>=2.7.0
